Generating circles breakthrough
A long method but for now i have 3 cirlces which have random position
and size. It is just a matter of copying and pasting parts of my code to
now make 20 circles. Im sure theres an easier way to do this but for now
i am pretty happy with the progress